+++ client/src/reducers/index.ts
@@ -1,10 +1,10 @@
 import { combineReducers } from "redux";
-import itemReducer from './itemReducer';
 import errorReducer from './errorReducer';
 import authReducer from './authReducer';
+import listReducer from './listReducer';
 
 export default combineReducers({
-    shoppingList: itemReducer,
+    shoppingList: listReducer,
     auth: authReducer,
     error: errorReducer
 })
